Fix floating crashing due to useless aliasing
For subqueries we always create an expressions that corresponds to
"thing AS column_title" in the select list. When a simple column is
selected this ends up as "SELECT column AS column". However, from other
(floating in particular) places we might create a "SELECT column", which
would be treated as not equal to "SELECT column AS column" when
compiling the final list of db_columns.