inline extract pattern compilation
It wasn't clear what "precompile functions" does. After digging a bit, it turns out it only compiled a string into a regex. Therefore, I decided to move this to the compiler module, and clean it up a bit, so it relies on lenses, and therefore doesn't need to do a manual recursive walk.