Only support `quarter` for MongoDB version >= 3.2
This forces the query to be emulated for older versions
of MongoDB, which arguably is not nice. The alternative
is unfortunately also rather unsatisfactory:

We could write an alternative projection for 3.0 that
relies on some `$mod` magic to simulate truncation.
This will however produce a floating point value rather
than an integer. To clean this up, we would need
additional post-processing magic in the cloak. Very messy.