Tweak constants in noise layer integration tests
These tests are sensitive to the exact combination of the constants
appearing, seeds set and the algorithm used in Anonymizer. This may make
maintaining them quite painful but I think that it's still better than
having no tests at all that touch this functionality... I guess the
solution is to not have too many of them.