Make SD per noise layer constant
It seems there is a possible attack that results in being able to strip
away noise layers if the SD per layer depends on the number of layers.
The attack assumes that constants are included in noise layers for <>,
see https://github.com/Aircloak/aircloak/issues/1278 for details on
that. In broad strokes what you do is:

* Create a query with two layers - the one you want to strip away and a
  chaff clause like `age <> 1000`. At the very least your query will
  have uid noise added that can also be targeted.
* Vary the constant in the chaff clause to average out that noise layer
  - that's an expected feature of noise layers for <> clauses, but here
  it allows you to obtain two samples for the same query with only the
  targeted noise layer(s). The samples will only differ by the factor by
  which the noise from the targeted layer(s) was multiplied.
* You should be able to obtain the exact result of the query by solving
  the system of equations:

  actual_value + 0.71 * target_noise = result1
  actual_value + 0.58 * target_noise = result2