refactor table columns representation in the datasource
So far, we've been using `{name, type}` to represent each column. With pending changes, we need to support "hidden" columns - i.e. uids which we implicitly select in subqueries, but which are not visible to the end user. To prepare for this, I've changed the representation of a column into a map with `:name` and `:type` fields.