Query state transition from error to complete illegal
I have seen a bug where a query ran out of memory.
It didn't have any rows or columns but contained the
OOM error message. All the same the query was in the
query state of "completed", which caused the web interface
to crash as the query it tried to render was incomplete.

I have not been able to reproduce this locally.
My current hunch is that there is room for some
race condition somewhere, which has the query state
be sent after the error has been recorded returned?

Either way it should never be legal to move away from
a terminal state, so that is what I am fixing here.
I hope that in turn means the error does not reoccur.