Add a try/rescue when logging a query
The queries we log like this might have some things that are not allowed
in queries that would be sent to postgres. At the very least Regex
constants can appear in them, so using the postgres dialect doesn't work
out of the box. On the other hand I don't want to add support for
Regexen to the postgres dialect just for the sake of logging, so that it
doesn't end up doing something on accident.