Inconsistent anonymizer between count and count noise
With count(*) and count_noise(*) (amongst others)
now returning `nil` when there is insufficient data,
it became very obvious that the anonymizers used by
count(*) and count_noise(*) were not the same.
Very frequently one of count(*) and count_noise(*)
would produce a nil value, but not the other, which
made no sense.

This pull-request ensures that count_noise(*), like
count(*) also becomes a starred anonymizer.