Ban date_trunc on dates
Postgres does a weird implicit cast involving timezones in this case. As
a result the fields are not zeroed as one might expect - for example
(depending on timezone) - the result of doing `date_trunc("day", ...)`
might have a 23 in the hour field. Because of that I think it's better
if the user has to do an explicit cast to datetime.