Update builder dependencies
I added a new dependency needed to compile offline
versions of our docs.

The jq dependency doesn't seem to be needed?
Removing it doesn't impact (as far as I can tell)
the building of the container. I therefore removed it.

Since it was unclear where the jq dependency was used
I also introduced a listing of why dependencies are
included to simplify keepign the docker files slim.