simplify chunk format
In previous commits, air was transforming chunks into buckets, while the client (browser) was counting rows. This required some extra complexity in air, so this commit simplifies things by moving to completely chunk-based interface. On first fetch, the first chunk is sent to the browser. If the client needs more rows, it will ask for the next chunk.