Raise thresholds for memory checks
After including cache as available memory as well I changed the
threshold to more aggressive values to avoid queries accidentally
killing the cloak.

It has been an iterative process of gradual adjustments. I am not quite
sure these values are right, but they are the least aggressive values I
could find that didn't allow me to kill the cloak with queries that are
memory intensive.

I tested a couple scenarios:

- Lots of smaller queries that were constantly issued along with a mix
  of queries consuming a lot of memory. The hypothesis being that the
  cloak would accidentally kill the constant short queries instead of
  memory hungry ones, hence dying.
- Lots of parallel queries all consuming lots of memory.

I am now unable to kill the cloak because of running out of memory.
I can still kill it by overloading the air <---> cloak communications
and have everything time out and thereby the cloak shutting down, but
that is an altogether different matter.