clear unused deps from travis cache
Yes, this is very ugly, but the alternative is to remove the master cache, and the branch cache, and prevent anyone from merging to master until this PR is merged. That sucks, so I think this is better, and I'll remove this code in a separate PR, once the cache is regenerated.