turn result sender into a plain module
Previously, we started one result sender process for each piece of information we wanted to report back. This led to a possibility of reports not being sent in the proper order. Also, it increased the pressure on the air socket, as a single query could quickly generate multiple concurrent reporters.