limit concurrency amount for query operations
If things happen in bursts, e.g. we issue a bunch of queries at once, or a bunch of results/status reports arrive at once, handlers start to trip over each other, due to many concurrent database operations. This is a simple fix to limit the amount of concurrent operations.

Not that this doesn't limit the amount of simulatenous queries. We just make sure that at any point in time, we're not starting or storing more than five queries.

This is a fairly cheap trick, and probably won't work properly for higher loads. However, I think it can get us fairly far, without needing more complex approaches. A better solution would be to have some "demand-driven" technique, like GenStage does. This would however require more complexity, and complicate the communication protocol, so I didn't want to do that now.