reduce chattines when querying for dead queries
So far, when determining the query status, we contacted each connected cloak once per each query. This can lead to a lot of chattines when there are many simultaneous queries. In this commit, the strategy is changed. We ask each connected cloak for the list of currently running queries. Therefore, when cleaning up, we need to issue only one request to each connected cloak, regardless of the number of running queries.