Use 4+ digits worth of reals data in seed
This makes the seed dependent on the data, yet masks floating point
inaccuracies for differences between 32 and 64 bit floating point
values.

From Wikipedia [1] I gather that:

> All integers with 6 or fewer significant decimal digits can be
> converted to an IEEE 754 floating-point value without loss of precision

This should indicate that 6 digits should yield a safe cutoff point for
floating point values. In practise however this does not seem to be the
case as anything more than what is currently done (4 + rounding of the
5th digit) produces variable seeds. (So far tested with `real` and
`double precision` in Postgres).

Also tested with data generated with only a limited amount of digits
(i.e. truncated to say 2 digits).

Furthermore, since we are using the column value in the seed, rather
than the result of an analyst initiated computation (say the analyst
selecting `n / 10^x` for some x in order to alter the precision)
the seed value is safe against malicious analyst fiddling.

1: https://en.wikipedia.org/wiki/Single-precision_floating-point_format