Darn complex adjustment of floats for seed
This now ensures that we only retain a certain number
of digits in floating point numbers that are used in
a seed, but also makes it so that similar looking
numbers still produce different values in the final
seed.

This got way more complex than I would have liked it.