Allow setting custom text encoding in data source config
This is needed in order to ensure that we can run against
arbitrarily encoded SQL Server databases.

Our own SQL Server instances have been unicode so far,
whereas the default for SQL Server databases on Azure
is Latin1.

The encoding of a SQL Server database cannot be changed
after the fact, so we need to be able to accommodate other
encodings too.