consolidate dialect polymorphism
The current code was somewhat convoluted. We had hardcoded mapping of a dialect to module in the Support module, and then we had a bunch of sprinkled conditionals in various places in the code.

This commmit changes the dialect field to represent a callback module, and moves all the conditionals into the dialect specific module.