Detect new, removed, and changed data sources in cloak
The operation is quite efficient for adding and changing
data sources, only the altered data source being checked.

On removal a very lazy approach is taken, reinitializing
all data sources, rather than trying to figure out which
one it was that was removed. For data sources that are slow
to initialize (like MongoDB), the cost can be quite high
as a result. The alternative would be to retain the file
name that a data source definition originated from in
the data source structure. Since removal (or data source
changes for that matter) is thought to be a low frequency
event, I decided in favour of keeping the data source
structure clean instead.