Mark distinct tests as pending
Turns out we have a long-standing bug when using distinct. When
something gets floated through the distinct it will include the floated
data in the distinct, causing the wrong thing to be "distincted". These
tests failing now is just a result of producing noise layers in more
cases.

The interaction between distinct and floating is not yet specified on
the conceptual level, so I'm leaving these tests for when that
functionality is implemented.