Make noise layers for columns from subqueries unclear
This is a simplification. We could detect the case when an unprocessed
column is selected in the subquery, but it doesn't seem that frequent
of a case to me, and the code is already somewhat complicated. I'm
leaving it for later since the whole clear condition idea is really just
an optimization in the context of equality conditions.