Removes `emulated` flag from function definitions
The `emulated` flag was a hack, initially used for handling
`extract_match|es`, that was overlapping with the responsability
of the emulator. Furthermore, the checks were buggy since some functions
that needed emulation could have been missed in the `WHERE` clause.
A new `supports_function?` callback is added to drivers which should
report if a function needs to be emulated or not.
This change also now allows the possibility of offloading some
row splitter functions to the database.