Support for ingesting data in parallel
Row grouping is separated from the rest of the aggregation procedure
and a new module is added that can group the data using multiple processes.
Once the entire data is consumed, merging of groups is performed and the
rest of the aggregation process is resumed sequentially.
The number of ingestor processes can be configured per data-source.
For very large data sets, the performance gain is significant.