Alter median aggregate
Closes #2054

The change is as follows:
- Drops noisy top-count outliers above and below
- Returns nil when there aren't sufficient users to
  drop values above and below

The changes required increasing the number of data points in
a number of tests to get sufficient data to perform the median
anonymization. This in turn led to changes in unrelated tests.

Since no changes had been made to other aggregates, the values
returned by the tests were mostly taken at face value and used
in the updated tests for the un-altered aggregates.