Include full expression when detecting implicit range
This makes it much easier to later find if this expression is allowed in
a uniform manner. When only the column to which the trunc/round/etc.
function is included in the Range object there is not enough information
to figure this out, because what also matters is that the result of the
trunc/round/etc. doesn't undergo any transformations before being used
in a condition.