Classify bucket as a discontinuous function
It isn't directly used as a discontinuous function
in the type checker as it gets rewritten into
basic operations as part of a pre-processing step.
However: if function classifications are part of
the function definitions, bucket has to be listed
as such too. Otherwise other subsequent users of
the classifications will get incomplete/wrong results.