Include min/max/count in all noise layers
This means that for noise layers that aren't affected by any aggregation
(for example in `SELECT COUNT(*) FROM table WHERE x = 3`) we will use
min=3, max=3, count=1. This has the benefit of being more uniform in the
code, but also, crucially, avoids providing the analyst with an extra
sample when it's possible to build queries with basically the same
semantics but where one of them uses floating and the other doesn't. For
details about that see https://github.com/Aircloak/aircloak/issues/2082.