Add regular noise for negative conditions on UID column
Previously queries adding negative filters on the UID-columns would not
get any noise values for these conditions. This resulted in the ability
to write query pairs like:

  SELECT count(*)
  FROM table

and

  SELECT count(*)
  FROM table
  WHERE uid <> 'victim'

and get a definitive knowledge about whether or not the victim was in
the dataset due to the change in the output noise.

------------------------------------------------------------------------

This fix does not include a fix for NOT ILIKE. I couldn't get that to work.