Simplify uid test for ILIKE 'trivial-pattern'
I am not sure how the test passed before, considering
we have rewrite rules that convert `col ILIKE 'Trivial-Pattern'`
into `lower(col) = 'trivial-pattern'`, but explicitly comparing
for that makes the test pass...

Not really sure this isn't hiding a bug, or shows what we
need to show at this place. We have one other test in place
to ensure that unclear patterns get the noise layers the expect,
so my hope is that the concrete contents of the noise layers are
checked sufficiently there.