Adds initial set of TeamBank queries
These questions have been shipped back and forth during support
emails with Bianca from TeamBank. There is one mismatch between
the queries as listed here and the ones sent to Bianca, and that
is that the BLZ columns are encoded as text in the data source
config in these tests, whereas they are currently encoded as
numbers in the TeamBank data source configuration. This will be
altered in future versions of the TeamBank installation.

This commit also includes a query which is currently forbidden,
but should be allowed as soon as we can make it so.

Some of these queries we have advertised to them in the past as
being of great benefit...

We should strive to re-enable them.