rename cloak ci image
There are some collisions with the older version (current master), which cause the image from this version to be deleted, and leads to failing tests. This is a simple trick which will allow both versions to coexist.

Once this is merged to master, the problem should be solved, because this version is less aggressive about removing images.