support per commit commands definitions
This will allow us to keep the test/compile commands in each commit, and thus avoid frequent CI version changes, and a bunch of conditionals.

Notice that this PR also raises version requirement for cloak_tests to 3. That's fine, since 2 is only on master, and cloak tests are currently experimental, so as soon as this is merged, we'll be running cloak tests on all subsequent PRs.