Generate no noise for uid IS NULL
This seems correct so long as we don't generate noise for uid IS NOT
NULL. However, the real reason I'm doing this is as a workaround for an
issue with floating user ids in queries with joins I'm having. The issue
can be reproduced by changing the condition in
test/cloak/query/join_test.exs:188 from `WHERE car.user_id IS NULL` to
`WHERE car.user_id <> 'whatever'`. The conditions should be equivalent,
however the second one crashes, because a row_index wasn't set. I lost
quite a lot of time trying to resolve this, so I'm leaving it for
another PR.