Optimize noise layers accumulation
It is a lot more memory efficient if we compute the hashes during the
data ingestion phase and we then store, copy and merge sets of hashes.
It is also more efficient to store the hashes as 58-bit integers,
as that is the size of the RNG state.
Speed is also improved by making sure that the noise layer values
are computes as an iolist.