Drop query execution reports in Air
As there are no other messages being sent from Air to Central at this
point. A lot of general Air<-->Central RPC tests had to be removed.
They relied on the query execution messages being sent and tracked
in order to validate the correct behaviour. For 18.3 this functionality
will be brought back (albeit in with somewhat tweaked message content).

Closes #2532