Truncate cloak log entries sent to Air to second
This makes timing attacks harder as the resulting timestamps
don't give as much detail about how long the cloak spent in a certain stage.

The truncate/2 method added to Cloak.Time was somewhat modelled after
the truncate/2 method in Elixir Time added in 1.6. Unfortunately (due to
the timestamp tuple used by the logger) this function couldn't be used
out of the box.

I contemplated adding this in the LogCollector module itself, but opted
for adding it at the latest possible state in the log progressing chain
to make sure the same truncation treatment is applied should there later
be other sources of log messages.

Closes #2601