Add an endpoint for authentication checks
It returns very much a boolean style response.
A 200 OK if the user is authenticated, and a 401
otherwise. This is useful as a checking endpoint
when nginx is used as a proxy for some other backend
service which itself should be authenticated.
nginx can then be instructed to perform authentication
checks against the app first, while all the hard
word of actually performing the proxying itself is
left to nginx.