Hack around emulated subqueries including user_id
It seems like user id is always assumed to be required in
`Query.required_expressions`, but that requirement is not in fact enforced
for `:standard` queries in `Specification.auto_select_uid_column`. This
makes it possible to create a query where the subquery does not include
a user id in the produced rows, but the outer query still expects it,
even though it's not used for anything, resulting in a crash in
`Selector.do_pick_db_columns`.

I tried removing user id from `Query.required_expressions` but that also
didn't seem to work for this case.