Reduce concurrency
The cluster mechanism seems to start somewhat
heavy weight threads that all run a copy of R?
I reduced the number of threads to an even multiple
of the number of queries we are currently running
(5 threads for 15 queries). This way we get better
and more even reuse of threads.

Ideally (but that is for a future change) I would
have per thread database connections that remain
open for the lifetime of the thread, to avoid
new connections having to be established for each
query that is run.