Add index to query_state in query table
On the activity monitor admin page we load all currently running
queries on page load to prepopulate the page with active queries.

For systems that have had a lot of queries run, the database query
to load these active queries was prohibitively expensive.
It includes a query condition on the query_state, excluding all
the finished states.

Since there was no index on the column, it seems a full table scan
had to be completed. On our demo host, the query would take 12.14s
of the total 12.17s page load time.

This is a first test to see if we can improve the page performance.
If the index does not help, then changing the query might be another
option, or alternatively not at all loading the currently running
queries, and instead only show those that start after the page
has been loaded.