refactor propagation of datasources to air
This refactoring prepares the terrain for upcoming changes, where we'll periodically recompute datasource info, and send it over the wire if it changes.

As an added bonus, this change breaks the coupled dependency between datasource and airsocket, and allows us to reuse the change publish mechanism in other places, such as the isolators cache.