Refactor fetching a list of queries
This should be a more idiomatic approach. An unfortunate side-effect is
that it's possible for data_source to be nil in some cases, and I had to
work around that in the view. I don't want to mess with that problem
now, as adding a constraint that forids it might have some consequences
in other parts of the code.