Remove test which no longer applies
This one is a bit tricky... error's are still reported
via the audit log when the system fails to delete a user,
however the reason that previously caused a user not to
be deleted (and the one which was tested for) no longer
applies as we immediately catch if a user can be deleted
or not due to first disabling the user. Lacking a good way
of testing the replacement functionality here, I instead
remove the test.