Use a custom flooring function in FixAlign
For float numbers the value of `a / b` might be slightly under the true
value. In that case `Float.floor(a / b) * b` will be `a - b` instead of
`a`. For FixAlign this resulted in sometimes aligning intervals that
didn't need it. This fixes the problem, at the cost of sometimes
aligning an interval that is very close to an aligned interval
incorrectly. That is for an interval `{a - epsilon, a + b}`, and
sufficiently small `epsilon` the aligned interval will be `{a, a + b}`
and won't contain `a - epsilon`.