Start reworking fuzzer
I want to make it so that most fuzzer queries can be ran against a data
source as opposed to failing at compile time. I think some large changes
must happen to that end:

1. Generation of some structures must go bottom-up. In particular
  subqueries for a given query need to be figured out, before that query
  can be generated.
2. Generation of some other structures must go top-down. For example
  some expressions are forbidden in subqueries, so we would like to know
  if we're in the context of a top-level query or subquery when
  generating those.
3. Currently the fuzzer either generates incredibly complex queries with
  many subqueries and very complex expressions or very simple queries. I
  want to take direct control of the complexity of what's being
  generated such that the amount of complexity "allowed" is "invested"
  in various ways - either building more complex expressions or a more
  complex overall query structure. This should make the fuzzer generate
  queries of medium complexity, making it more likely that such a query
  will both be useful and able to run.

Achieving both of these with stream_data is somewhat tricky. The size
parameter which is supposed to guide generator complexity is difficult to
properly manage. It's also somewhat cumbersome to generate data when
both a top-down and bottom-up approach are needed. Because of this I'm
trying a hand-crafted approach.