build release in container
This change removes long and complex docker builds. The reason for that is to improve the behaviour on the CI server. Docker builds have issues when executed concurrently. They might take significantly longer, and they might even fail occasionally. Therefore we want to run one docker build at a time. However, this means that docker builds in general shouldn't take a long time. With this change, we do most of the heavy work in the docker container, which means that docker builds are usually going to be short.