support docker build phases
This change allows the component to specify multiple phases which enter different queue. This is helpful for system test, where most of the time is spent outside of docker build. With this support, we can improve throughput, by keeping system test outside of docker_build queue while it's not in fact using docker build.