Stop adding quoting in (R)ODBC table discovery
It doesn't work for tables with a schema. For example in SAP IQ you can
specify a table as `user.table`, especially if many users have a table
or view with the same name. The preferred solution will be for the
administrator to add the needed quoting in the config file.