Make the RODBC port driver async
Blocking the main emulator thread for long periods of time hangs the
Erlang VM. Making the operations async, we avoid that and instead block
one of the threads in the async thread pool.

This also eliminates the possibility of pooling RODBC connections (they now
need to be used from the owning process only).

The change still doesn't solve the issue completely for long-running operations,
as it seems there are many drivers using the async thread pool, which make
the cloak unresponsive until the long-running operation completes.