Use the JSON storage for drill compliance
There are problems with both using mongo and basically any other
storage. For mongo the problem is that the data is nested, and gets
flattened in a weird way once drill queries it. For postgres, mysql,
etc. there is a long-standing bug that makes it so that queries with
`LIMIT 0` or `WHERE 1 = 0` crash (https://issues.apache.org/jira/browse/DRILL-4984?page=com.atlassian.jira.plugin.system.issuetabpanels%3Acomment-tabpanel&focusedCommentId=15622085#comment-15622085).
We use such queries to discover table columns (and it seems
`:odbc.describe_table` does too), so we never get past loading tables in
that scenario.