Fix crash for noise layers on expressions with UID
Because UIDs have special handling in many places it's somewhat
difficult to instruct the query engine to select them. This makes it so
that noise layers are built using the top-level UID in place of any
other UID that might appear in expressions. It was already mostly the
case, but only for UIDs we mixed into the noise layers ourselves. Now
UIDs that end up in noise layers because of how an expression looks are
also given this treatment.