Improve result generation
Instead of creating a deeply nested structure of test results
that is somewhat annoying to further process, I am not producing
a long list of query results instead, which can be sliced and diced
to our hearts content in order to get at the stats we desire.

In addition to tracking the real vs anonymized counts, I now also
precompute the error and the relative error, which makes producing
stats a breeze.