Dedup negative conditions for shadow check
This way, if the analyst includes the same condition twice by accident
we won't count it against the limit of 2 with rare values. The
deduplication is limited to simple `column <> constant` conditions,
because I want the mechanism to be as simple as possible to minimize the
potential for bugs leading to privacy violations.