Add failing tests to validate hypothesis of missing normalization
We normalize queries in the query compiler, replacing such things
as DISTINCT with GROUP BY's and other similar things.
This normalization only takes place in the query provided by the
analyst, and not in queries that are provided either as a view
or that are defined as queries in the data source configuration.