improve connect workflow
Previously we connected from `handle_continue` which required a slightly confusing error reporting logic, and more importantly there was a slight chance of a connection crashing on connect error.

With this change we'll connect on first usage, which in practice happens immediately after the connection is checked out, so there are no particular performance implications.