improve stop_query
The tests have revealed some strange behaviour and guarantees about the way queries are stopped. For example, if no cloak is connected, we refuse to stop the query, but if only a wrong cloak is connected, we'll stop the query. Likewise, if we succeed to stop the query on the right cloak, and then a disconnect happens on the wrong cloak, we fail. Finally, stopping was done in a synchronous way.

With this commit, the guarantees are lowered, but at least it's simpler to reason about the code. We'll try to stop the queries on the connected cloaks asynchronously, and put the query into the cancelled state regardless of the outcome.