Put user logins in a separate table
This is in preparation for adding app passwords for accessing the PSQL
interface. To avoid timing attacks the login process needs to be
uniform, regardless of number of app passwords configured for the user.
To achieve that, the app passwords will come in the form of (login,
password) pairs and be stored in the same table as the main password.
The table will be queried like the user table is today, and at the end,
assuming the login information is found to be valid, the user associated
with the given login will be returned. This will of course be scoped by
login type for PSQL/regular access.