Create compliance tables in Oracle
I'm using the jamdb_oracle driver which doesn't seem very reputable.
This is because we're in a bit of a hurry and connecting via ODBC seems
to be full of problems, in particular our RODBC driver doesn't seem
compatible at the moment. The driver will need to be replaced sooner
rather than later.