Update to latest pre-16 version of react and update CSS transition group
The version we were using of CSS transition group was deprecated.
Now we are up to date with the latest version of this dep.

React is now pretty much exactly one year old. It would be good to
get up to version 16.X at some point too, but we have other dependencies
which are causing problems