Use PureComponent to avoid re-rendering as many times
This seems to help push the performance a bit up from my experience
in the web interface. However, we are still producing way too much
garbage. The GC is kicking in really aggressively, and pausing the
execution for 250ms at a time, dropping the framerate from 60fps
to about 0.5fps... no wonder if feels laggy.

Oh, also I could not figure out how the CRSF token made it to the
cancel query button... As far as I could tell it was never passed
to the component. Now it is.