Relax SAMPLE_USERS compliance test
I decided to use a different hashing algorithm for oracle (ORA_HASH
whatever that does, instead of MD5). The reason for that is that MD5
seems supported in different ways in different versions of oracle, with
quite a bit of extra complexity. Because of this, the results for
SAMPLE_USERS can vary quite a lot between oracle and the other
datasources.