Store Postgres performance db data in a volume
This way we don't end up clogging up the docker overlay
storage system (which doesn't get properly cleared).

To ensure each test-run has a clean database it's cleared
between test runs.

Additionally I have made the output from the tests
somewhat dirtier. The existing cleanup wasn't correct
because of other cloak changes, and it seems this is
likely to be come the case again in the future.
Rather than chase the differing output I think it's
OK to have to do some manual cleanup.

It now also includes metadata about the size of
the database being tested itself which is useful for making
better assesments on the performance.