reduce blocking in Streamer
The streamer process was blocking, waiting for the ack from the driver (either an error or success). This means that if the driver waits for too long (which we've seen happen with Oracle), we actually get a timeout.

This commit moves error/success reporting outside of the blocking loop to the processing pipeline. As a result, the streamer now waits for the ack for more than 5 seconds, so we don't get false connection timeout errors.