Improve cast from text generation in fuzzer
Cast from text would often fail, because the input (often a random text
constant) would not be properly formatted for the given output type. To
work around this, and still generate various cast I always generate
expressions such as `CAST(CAST(foo AS text) AS type)` where `foo` is of
type `type`. This does not test casting so fully, but should still
exercise it, and I decided it's worth it to get rid of false positives.