Refactor query generation
By wrapping the complexity and tables in a single context map it will be
easy to add more information to that context. For example when
generating expressions it would be useful to know what type of conditino
(<>, =, etc.) the generation is taking place in, to be able to judge
which kinds of expressions are allowed.