Remove appsignal
Turns out our reporting to appsignal has not been working for some time
now, most likely since the upgrade to Elixir 1.7. Because we got no
information that this was happening and restoring it to working order
would involve some amount of work, we decided to remove appsignal for
now, possibly to be replaced with another monitoring service later on.