add application-level connection limiter
The ranch limiter is soft, which means it can be exceeded quite easily. For example, even with max_connections set to 1, I was able to establish a few thousand connections on my machine. The application-level limiter imposes a hard limit, which can ensure that even if ranch limit is tripped, we don't end up establishing the PostgreSQL connection. In addition, the application level limiter drops the connection instead of queuing it indefinitely.