Sign the revokable token's id instead of payload
What I wanted to achieve with this is to have a cryptographically strong
mechanism preventing someone from accessing a random token (for example
by guessing it). It doesn't seem UUID generation guarantees that
(although it should be pretty good). Phoenix.Token.sign on the other
hand exists for expressly the purpose.