remove db_select from the struct
There's no need to keep this in the struct as it can be always derived from the query. This requires adding a new function to the behaviour contract, but we'd need to do this anyway, because not all drivers are necessarily based on SqlBuilder, so the code in `MaterializedView` shouldn't invoke `SqlBuilder` in any case.