Remove guardian
Now, that we store sessions in RevokableTokens, guardian is not doing
anything anymore. To be honest, I don't think it was giving us that much
before - we could have just been using `Phoenix.Token` that whole time.
We certainly didn't use most of the features in guardian, we simply
wanted to set a user_id in a cookie in a secure manner.