Account for race condition upon page load
The problem could arise that a table creation completes between
the page has been rendered on the server and the web socket connection
has been established with the backend. In such a case the web client
would never receive information that the analyst table completed creation,
and would be left with an indefinitely spinning (and unusable) analyst
table.