Relax FixAlign property
For intervals that are very nearly aligned (at around the level of
floating-point accuracy), like {0, 500000.000005} it doesn't make that
much sense to enforce that they be aligned to a strictly bigger number
and allowing {0, 500000} to be the result is fine. To put it another way
- x <= y for floats should really be x <= y + delta, where delta is a
small number relative to x and y, due to floating point imprecision.