send air name to cloak on connect
This will reduce the need to pass the air_name around in various places. Most notably, we don't need to support air_name as a parameter to query execution.

The downside is that with this design we can't support the scenario where a single cloak is connected to multiple airs. However, supporting this scenario will require more work anyway, so I believe this approach is fine for the moment.