Handle noise layers of selected fields in analyst tables
When a field is selected in the created analyst table we don't reselect
it under the `__ac_nlc_*` alias just as in regular queries. However,
when querying that table with no usage of that field except for noise
layers, it would get assigned an `__alc_nlc_*` alias and fail to be
found in the created analyst table.