refactor drop table interface in cloak
This will allow us to drop multiple tables when deleting a user.

Also, the constraints are relaxed, and the cloak doesn't return an error if the table can't be deleted in the database. This is by design, b/c we don't want to prevent deleting the table from the air if something is off in the cloak. Database admin can always cleanup unused table manually in such cases.